requirements.txt
# requirements.txt
Flask==3.1.0
gunicorn==23.0.0
Pillow==11.1.0
psycopg2-binary==2.9.10
pypdf==5.4.0
reportlab==4.3.1
python-dotenv==1.0.1
